DCP00{0[0-9],1[0-2]}: Add voting results.
This adds a new voting results subsection to the deployment section of
all DCPs to document the historical blocks at which the voting started
as well as the locked in and activation heights.

It also clarifies that implementations may optionally optimize their
enforcement activation logic accordingly for those historical activation
points as opposed to tallying historical votes in the applicable DCPs.